Big push
This includes a lot of basic functionality that is necessary. Most metadata is
supported, you can edit metadata and thumbnails. You can delete and add
libraries and the thumbnails get cleaned up as well. Default values work mostly,
 except that upon saving a video the default value will be taken as the actual
value, even if it was not edited.

This also adds in database migrations so I can avoid having to resort back to
hibernate for schema generation and updating.