Add functionality to search field in client
This fills the existing search field with life as it now actually does a search
using the grammar and parser we previously created. The searching currently
works by making a subselect for each partial query and checking if the video
matches that query. This is because we have a list of values we need to check
against, instead of simple properties that we can access on the object.