Create sorting view
This is an experiment I'm trying for a better way to initially index the videos
since currently one had to manually select all of them and enter the same
information over and over. Instead, this allows predefining a set of values for
a specific kind of metadata and then goes through each video that doesn't have
any values where the user now only needs to selects one of their predefined
collections and they get applied.