Use video.js as a video player and implement video preview thumbnails
This makes use of video.js for the video player instead of the default html5
video player. Additionally, this adds thumbnails for when hovering over a
timestamp in the video to preview that part of the video. This is implemented by
 showing the same video from the hovered position as a preview and thus doesn't
require another set of preview images as well as enabling an animated preview
where the video is played instead of just a still image. The downside is that
instead of a simple image, a normal video needs to be loaded. Sadly, the browser
 doesn't reuse the same video thus possibly causing two full video downloads.
This is often not really an issue since we're talking same network here, but on
large videos (500MB+) this can become less responsive.