Redo thumbnail plugin and fix thumbnails after video change
Recently it was fixed that the scrubbing thumbnails did not get updated at all
when the video was switched. This however had an issue in that it created a new
div above the existing div, thus stacking multiple videos on top of each other,
which obviously was not great. This now redoes the whole thumbnail plugin to be
a proper class and thus be able to re-use the created elements which in turn
allows us to just change the src url of the thumbnail-video.