Encode metadata options and values in json object
SQLite provides us with the option to use JSON functions in queries and therefor
 can use json to store the data. We do that because that way it's much easier
given the different kinds of data we're storing. We might lose a little bit of
performance to parsing/serializing the objects to/from json, but we can now
store much more complex objects and query them pretty well. This sadly required
us to bundle the value in some container object which now requires another step
to access on the client side.

The only side that's still a little ugly is the query building as we need to now
 build raw native queries and currently don't do anything to escape the values
allowing easy sql injection. This isn't that much of an issue given the data
that is stored combined with the fact that this should not be accessible to the
internet. It should still be dealt with at some point for sure.