Improve performance of adding new metadata
When adding new metadata, we previously loaded every video, went through all of
its metadata values and checked if the metadata definition was used. If it
didn't we added it. This resulted in a lot of queries, since we loaded a list
for each video as well as each video itself.

We can cut down the number of queries by doing two things. First, we avoid
loading the values for each video by just assuming that no video has the
metadata value yet (because we just created the metadata). Thus, we no longer
load the values of each metadata for each video. Secondly, we enhance the
relationship of video to a metadata value by adding a back-reference. This
normal is not a good idea, however, it makes sure that hibernate doesn't need to
 first do an insert and then an update, again avoiding another statement.

It's still rather slow, since we need to insert so many entries in a big
collection. I presume it could be made better using a stored procedure to