Add sorting and infinite scrolling
If we have a big library (>1k videos), vue will start to struggle quite a lot
when loading all videos at once. Moreover, the data that has to be loaded is
also not that small making it even more intensive. By doing using an infinite
scroll approach we can make it as easy as possible for the user and still keep
the advantage of not loading everything at once. If the user scrolls for long,
we'll still run into a performance issue with vue as it cannot handle that many
components in the view so there's still a solution to be found for that. That
may require one more change that can now be done pretty easy as well (total
amount of videos).

This also adds in sorting and direction for displaying videos. Whenever the user
 searches, it will now order the results by the selected metadata in the given
direction. It currently doesn't automatically apply upon changing the property
or its direction, as this is a costly application and therefor should be
specifically called by the user. This may change in the future, however.
The implementation for this is a little ugly because spring doesn't provide
pagination for native queries as well as we have to implement our custom sorting
 for our properties as the values reside in the json structure. The performance
is also pretty good, so there's no trouble here. The only downside of the
current approach is that we do one query too much if we don't have any search
query, which is probably fine.