More progress for playlist
This includes a few changes that I can't bother to unwind and put into separate
commits.

One change is that we no longer return the full playlist, but instead just the
IDs of the videos from the playlist. This is to avoid loading all the videos
when sending just the playlist.

This also adds a button to add videos to the playlist which pops open a dialog
to add to an existing playlist or creat a new one.

The video player now also shows the queued up videos alongside the video and
automatically starts the next video if available.