Do not cache suggestion values but instead make query faster
The values were cached because the query was pretty slow and this it made more
sense to cache them in the client. This introduces the problem of inconsistency.
 We initially kept onto this behavior because suggestions should be possible for
 all values. That, however, isn't true; there are only two possibilities for
getting suggestions. First the tag type and second the suggestion option in a
text metadata. For most of the other cases it just doesn't make any sense or
isn't wanted in the first place.

Since we're now limited to just two cases, we can now easily optimize these two
queries and get great performance by avoiding the json transformation/parsing
before filtering.