Allow for disabling/enabling plugins at runtime
Currently, once a plugin is detected, it is activated. This makes usage somewhat
 easy by just placing the plugin file at the right place and it works. But
managing multiple plugins is cumbersome this way. Since the main idea for
configuration for vimax is to make things configurable on the fly in the UI,
this follows that sentiment.

Additionally, since we now also know when a plugin was added (because we don't
have a record for it yet) we can catch up on the changes it would've made to its
 metadata. Similarly this could happen once a plugin gets enabled again after
having been disabled some time ago. This is currently marked as a TODO in the
code because I'm not quite sure if that's a good idea or what the best way to
approach it are.