Detach videos to display from internal list of videos
Previously we had a list of videos that we maintained internally that just
contained all the videos that we needed. This list was also used for displaying
the videos on the video page. That's generally not an issue but here it also
included the order of the items they should be displayed in. As a result, once a
 video was loaded, it would not get displayed in the proper ordering.

To mitigate this, we now use a different list which is only there to have the
list of video ids it should display in order. This also means we no longer have
to clear our cache of videos we've loaded before.