Save owning plugin of metadata
Previously we only saved if a metadata attribute was specified by the
system, i.e. a plugin, or defined by the user. At the time we didn't
have an reference to plugins so this was the easiest thing to do. Now
that we also save the plugins in the database, we can link the two
together and thus allow for cascading to take place if necessary.

Sadly, sqlite doesn't allow renaming or deleting a column, which is
why we need to do those acrobatics to change the field. Additionally,
since we have foreign keys on the metadata table, we need to make sure
 to turn foreign keys off for the time being until we're done.

For some reason, the migration is required to run in mixed mode, which
 is why we need to enable the mixed mode migrations in the application
 settings.