Call plugin on enable for missed updated/created videos
Once a plugin gets enabled, it should be called for all the videos it
has missed in the the time it was disabled. Especially for plugins
that only have an onCreate handler, this is important because
otherwise there's no way to fill the videos with that information
apart from checking on every update - which is not ideal.

If a plugin does have an onUpdate handler we assume that it can handle
 unfilled plugins in here too so just use that instead of the onCreate
 and avoid possibly duplicating work.