Fix reordering of metadata
In the past when re-ordering metadata in the client, the server would get an
exception when saving because some display orders have clashed or other things
just weren't quite right. To prevent that from happening again the ordering code
 was moved into the server side which all elements will be looked at. Since we
might want to support moving elements to arbitrary positions, an algorithm was
used to figure which things need to be moved. The client currently only moves it
 by one in either direction but can be adapted to move them to arbitrary
positions.