Fix disabling all plugins on startup
On startup we check if there were any plugins previously present - and enabled -
 that are no longer present as a script and thus are unloaded. We automatically
disabled these on startup, however, this accidentally disabled all plugins on
startup. This was because we kept a copy in memory to avoid having to deal with
hibernate issues, but this copy didn't include the id. When checking if a plugin
 was loaded or not, we did use the ID causing every plugin to appear to not be
loaded and thus be disabled.