Make sure all videos have all metadata on startup
When initially starting, we'll now make sure that all the metadata is present on
 videos to hopefully avoid running into inconsistencies. The way this is
implemented is by getting the count of each metadata container for a video and
comparing that to the count of the registered metadata. This obviously doesn't
fix remaining values for metadata that has since been deleted.