Fix wrong assignment after editing bucket
When editing the value of the bucket and immediately after clicking on "Assign",
 it may assign the video to the old value (or null). The user would need to
first unfocus the input to make sure that the value is saved and then press the
button. Since the view is meant for fast entry, this is rather annoying and
therefor this now provides a workaround. When pressing assign, we don't
immediately assign the video, but instead wait 100ms to make sure the unfocus of
 the input gets triggered and only then assign the video.