Allow refreshing a plugin to update metadata on all videos
Plugins are supposed to be simple scripts or relatively simple. But even those
scripts change and could therefor provide other values for their metadata. And
instead of having to wait until a video gets updated to assign the new metadata
value, a user may now just refresh the plugin to update all videos at once. This
 obviously may take a while so a small update screen will be displayed.