Fix video not showing in sorting view anymore
Recently the player was moved to use the videos stored in the vuex store instead
 of the ones provided to it by properties. The sorting view used this behavior
and thus lo longer worked. It also had its own copy of a list of videos which
made the two approaches rather incompatible. This has now been reworked to only
keep a list of IDs of videos and load them into the store when used.
Additionally, we mirror the current video to the video player to be in sync.