Revamp storage of metadata values for videos
For each metadata, there was a row in the database which contained the value for
 that metadata for one video. This made sure that we had a row that had foreign
keys to the video and the metadata which ensured that once the metadata was
removed, it was also removed - same for the video.

Loading the metadata for a video was straight forward: select all the values
with a certain video ID. However, filtering was rather problematic. Since we
want to allow for and/or combinations, we had to do sub-selects for each term in
 the query. This _devastated_ the performance even with simple queries making it
 not really useful to use as it took too long to load the data you're looking
for.

This caused the redesign of how values are stored. SQLite performance with json1
 is great which allows just putting all json values into a single json column
for the video. It still makes sure that values get cleared up (because it's in
the video table) but we lose the reference to the metadata definitions.

Values are now stored like this:
```
{
    "1": {
        "meta-type": "BOOLEAN",
        "value": true
    },
    ...
}
```
We still need the "meta-type" for kotlin/java interop to allow for proper
deserialization.

Doing a query on this now doesn't require any sub-selects (only for default
values) and is thus tremendously fast (sub second respond times). Migration is
also rather straight forward and is doable without any data loss. In the
meantime, the query creator also uses a sql-injection safe prepared statement
instead just raw queries.

When the plugin api was adjusted, some peculiar things started happening.
Specifically updating a video through the api took a huge amount of queries to
facilitate. This seems to be because some dirty-checking is triggered as objects
 are passed back and forth from/to the script runtime. I took the opportunity to
 also change the plugin api to work with DTOs instead of the raw video. This
avoids the dirty-checking problem causing no unnecessary queries.

The only problem currently is that loading metadata for plugins (to access the
values of metadata not from itself) is not performant. Technically we just need
it for mapping between name and id, but the name could change at any time making
 it not possible to just load it once. We now cache the metadata when reload a
plugin, when updating a single video, nothing is cached. This also means that
ffmpeg and the like can be provided using a context instead of the bindings.