Improve performance of initial load of videos in client
The initial load of the videos in the client takes between 3 and 5 seconds which
 is arguably too long to the amount of data we're displaying. The server is now
not the bottleneck here making it only a client issue. One of the bigger issues
is committing each video separately into the store resulting in more work for
vue to diff the changes. By instead adding all videos at once, we already
increase the performance by a large margin.