Use IDs instead of the videos for video card and use v-show instead of v-if
In some cases we can use v-show instead of v-if since we don't have an else.

We currently iterate over the list of actual videos, which is rather wasteful.
Reason being that we don't actually use the values from the video in that case
as the card itself will load the necessary values. Instead, we could just
iterate over the IDs and lose nothing but get more performance.