Allow for editing multiple videos at once
This adds a dialog that allows the user to edit multiple videos at once.
Currently it's simply assigning the value set in the dialog to all videos if the
 value should be assigned (requires checkmark). This could be improved for
things like tag-values where instead it would just add the values if not
present.

This dialog will automatically open if you have more than one video selected and
 click the edit icon on one of the selected videos.

Closes #2.