Fix saving unchanged metadata element
Hibernate kept persisting metadata object even though they haven't changed. The
reason for this was because we used a custom type and for those hibernate just
does an `equals()` check if things are the same or not. We didn't provide an
`equals()` for the metadata options, causing it to always return false (because
it's pretty much guaranteed that it will use two different objects - though
they're the same). By providing an equals/hashCode and/or use data classes
hibernate now properly sees that there haven't been any changes.