Only set to null when metadata key is missing for video
On startup, we make sure that each video has keys for each metadata in its
metadata field. Additionally, we fill this field with the default value of the
metadata it's assigned to. This is not completely bad, however kind of defeats
the purpose of default values and the query executor also expects missing values
 to be marked as null so it'll automatically fall back to metadata assigned
default value.

Relates to #1