Fix illegal argument exception when on first start
When starting the application, each plugin will try to get or create their
metadata. Since it's the first start, no metadata has been created and the get
call will thus result in the exception. This however doesn't allow for the
fallback to happen (which is the create the metadata) as that only works with
null metadata. We change the behavior to just return null instead which allows
for creating of the missing metadata.