Configure complete build in gradle
This now allows a full build of the application to happen. First it builds the
frontend using `npm run build` and copies the resources and then creates a
fat-jar including all dependencies. This specifically does not use `bootJar`
provided by spring boot plugin as that has some issues, for some of which I
haven't found a solution:

1. Cannot load kotlin compiler. When trying to load the kotlin compiler, the
   class loader won't find the required classes. This can be worked around by
   telling spring boot to unpack the kotlin compiler jars before loading them.
2. Script runtime cannot load application classes. Once the compiler works and
   tries to load the scripts, it will fail again because the scripts use classes
   from the application. However, the classloader for the plugins doesn't seem
   to have access to them/doesn't know how to load them. I haven't found a
   solution to this issue which caused me to abandon it.

While the fat-jar approach does have its own set of issues, I was able to create
 a fully working artifact this way, which wasn't the case for the bootJar
approach. I'd be happy to go back but that's what we're stuck with for now.