Allow specifying external location for plugins
Currently, the plugins were only loaded from inside the jar - or rather the
classpath. This obviously isn't great for people who actually want to use this
thing and therefor requires allowing plugins to sit outside of the application.
This now adds a customizable folder location where a user can place a `*.kts`
script and it will be loaded.

Closes #7.