Return full list of result IDs when loading videos
Previously we used paging to only return X amount videos and dynamically load
the next videos after. This has two disadvantages: the query multiple times and
executed multiple times which costs time. Secondly, if videos change, the paging
 gets messed up, possibly missing videos or containing duplicates.

The were two options to implement this: either create something like a "cursor"
or similar to keep track of the searched results server side and advance on the
cached lists or just return all IDs to the client and let it cache all IDs.
Since we already only return the IDs, it's rather straight forward to do the
second option. Given the expected amount of elements is a few thousands, this
should be completely fine.