Provide option to run handler at a certain priority
Since a plugin might want to read values from other plugins' metadata, we need
to be able to specify when we want to be run. This was previously possibly by
just re-ordering the files, since the load order implicitly specified the
execution order. However, this is mostly just an implementation detail; this
might be changed in the future. Therefor, this provides a way to explicitly set
the priority.