Do not use video player store for non-watch views
The video player currently uses the vuex store to make it easier to keep track
of playlist data, video information and also make it easier to updates these
from somewhere else. However, this becomes problematic when there are multiple
usages of a video player, like we have here (watch view & sorting view).
Especially when switching between the two, as we'd now also need to make sure to
 update the current ID whenever we switch back and forth.

Instead, there'll now be a new player component, which is a much simpler version
 of the video player. It is not connected to the store and uses the video passed
 in by the props. It does not know what a playlist is and thus doesn't need to
manage skipping back and forth between videos in there.