Fix exception when first starting the application
When we initially start up, we create a black plugin information object which
contains more details in an embeddable. However, since we directly save it (so
it gets an ID) the embeddable will be set to null(!) as it's empty. While we
make sure it is there in the constructor using a default value, that will be
overwritten by hibernate. Instead, we need to manually make sure it is present
when it's needed.