Add checkin view
This adds a new video to the application called 'Check In'. While both the
"Check in" view as well as the "Sorting" view have the goal of assigning
metadata to videos fast they achieve this differently. "Sorting" allows you to
quickly sort videos by placing them in buckets, which means you assign the same
metadata in each video. "Check In" on the other hand is meant for assigning as
much metadata as possible for videos that have just been added. While one could
just easily go into the library view and manually edit the video, it doesn't
scale for when you have an enormous pile of newly arrived videos.