Implement Plex-like video player to replace current one
We previously had a separate page for playing videos and showing the current
play queue. This was annoying since there was no way to amend to the play once
you started watching as you'd need to leave the page and there was no way to
come back. Additionally, you couldn't let the video run while quickly glancing
through the library.

This now implements a player that works similar to the old one (scrubbing,
playlist display, scrubb preview) but more similar to the plex built in player.
For example, it can now be "collapsed", i.e. minimized, to allow for continuing
to use the application normally. As a result, the queue can now be appended to
while a video is playing.

The only downside for now is that opening a video in a new tab to play it is
currently now possible.

The old player will still be used, though only partially. It will stay in the
checkin and sorting view to make it possible for having a video collapsed while
sorting videos.