Fix lazy loading no longer working
Since we had to move the scrolling container somewhere else because of the new
player, this sadly impacted the lazy loading of videos. That still looked for
the scrolling of the whole window, which since no longer works. This sadly
requires us to move the scrolling container further down as we now need a
reference to it to apply the desired behavior.