Fix getting stuck after playing multiple videos
This is similar to a previous commit (d923a123), but it tackles the problem from
 the client side. You see, if we have a `<video src=... />` element inside a
`v-if`, that video will continue to buffer, even if it's detached from the dom.
Since it's not attached, the `src` attribute is not updated and thus is kept the
 same.

By manually setting the `src` to an empty string and then removing it, we make
sure that the browser stops the buffering of the current video and thus closes
the connection used for the request.

I'm not quite sure how that fixes the issue that a transaction is kept alive and
 not reset into the pool. Especially given that the logs indicate that all
transactions have finished, i.e. none is running and thus all should be
available in the pool again, yet it couldn't acquire one for a new request.
However closing the connection for the request seems to help here.