Increase database connection pool size
Initially, the pool size was set to 1 since I assumed there could only be one
simultaneous connection. However, this turns out to be false and multiple thread
 can in fact access sqlite db at the same time.

It doesn't have a huge impact since most actions are done in serial, but there
are some things that are indeed smoother. Most notably, the loading of
thumbnails as now multiple thumbnails can now be loaded at the same time.