Fix issues with ordering of metadata after deleting one
We currently assign an increasing number to each metadata to indicate the order
in which they are in. If you move the metadata around, we update its value as
well as the values of those metadata that are now or no longer before or after
you. However, when you delete metadata this left a hole in this chain, resulting
 in the client bugging out.

To remedy this, this now makes sure to update all those values that came after
the metadata that was just deleted. Unfortunately, this is not directly updated
to the client, thus a reload may still be necessary for now.