Delete missing videos on startup
If a video was deleted while the application wasn't running, we would never
notice this as deletions we only handled by the watch task. This resulted in
errors when trying to play this video. To avoid that, we check through all
videos on startup and remove all the ones that couldn't be found.