Fix not running job when trying to import previously deleted video
When a video got imported into the library, we create a spring batch job to
handle the generation of thumbnails and the like. So when a video gets deleted
and then re-added, spring won't run that job again because it uses the same
parameters. By adding a separate, mostly useless, parameter that is guaranteed
to not be the same, we bypass this blocker by spring and schedule a new job.