Add server side events to get notified on video removal
This lays the base for pushing updates to the client in case of certain events.
In this first iteration, we push if a video has been deleted on the server side
to avoid running into errors if the video no longer exists. While one could've
used WebSockets for this, server side events are way more appropriate for this
usecase as we'll only be pushing data to the client.

Additionally, this is also relatively straight forward compared to implement in
comparison to websockets. When a video gets deleted, we push the ID to all
clients that are currently listening. Additionally, we already also push a video
 if it got created, but it's not used in the client yet.

This did require setting the async timeout to `-1` to avoid the connection being
 closed after a certain time. A reconnect basically misses around 5 seconds of
information which would create a lot of work trying to avoid that. Also notice
that we send some data at the start of the emitter connection. This avoids the
client timing out because it doesn't know if the connection actually receives
any data. By sending a small "hello" message, the client sees that the stream is
 alive and will continue to listen.