Cut video names if they are too long in cards
Currently, if the name of a video is too long, it will be split into two (or
more) lines. This is bad for two reasons: first, since these cards are displayed
 as rows, the whole next row will shift down accordingly creating some wasted
space. Secondly, it makes it hard to "virtualize" this view as it's not trivial
to calculate the size of the element. By confining the name of the video to a
single line, this is made much easier as the height is now a bit more "fixed".
We're still some way off from this goal however.