Make tables for metadata even denser
This does two things: first, it removes the padding from elements in the
metadata value table. Second, it truncates multiple tags if they're too long.

Reason for this being is so that we can guarantee a specific height for the
table given the amount of metadata that is selected. This is important when
trying to virtualize the list of videos.