Make player handle better if video is not yet loaded
The player currently relies on the fact that the video has been loaded already
and that it can use all the available information. This however is not always
the case, especially with (soon) smart playlists. Therefore we make try to rely
more on the ID of the current video instead of the video information itself.