Update dependencies and replace mugen scroll with custom virtual grid
This updates many of the frontend dependencies to a more recent version. It also
 replaces the infinite loading in the client with a custom virtual scroll
component. While that component is not without issues, it comes closer to the
user experience I'm looking for. Given that the infinite loading results in
performance problems if you kept scrolling for a while, it does not work well
for large libraries. A virtual grid will serve any size of library and gives a
better representation of how large the library even is in the first place.