Use jobrunr instead of spring batch
We only really have one use case for a background task which is the importing of
 videos, which includes the generation of thumbnails and execution of plugin
event calls. Especially when importing a new library, the application should not
 lock up until imports are complete. The spring batch implementation is simply
too loaded and complex for something this simple. Additionally, for this use
cases it makes it more complicated than it had to be.

As a result, this uses a more simple solution: jobrunr. We still try to provide
the same guarantees that the spring batch version provided us.