Refactor selection values to be referenced by id
Selection values weren't in great shape so far; their values were referenced but
 could then also be changed in the metadata definition requiring an update on
videos that used that value. Except, it wasn't possible to differentiate in some
 cases between the user having replaced a property or just removed and added a
different one. As a result, this has been left out so far.

This refactor aims to resolve this situation by giving the selection values some
 identifier. This ID helps to make each value unique inside the metadata
definition. As a result, videos don't need to be updated on a rename of an
option. It is now also possible to reconstruct if an old value was just renamed
or removed and a new one was added.

The obvious downside of this is that it is no longer possible to get the name of
 the selected value without having the metadata definition since the value is
now just the ID. This is mostly important for plugins as they now have to adapt
to that.