Only load a certain amount of items in the play queue
To avoid loading a huge amount of videos when dealing with the play queue, we
now only show N amount of videos before and after the playing video. This is
similar to how plex handles it. For now, this is a static value of 20, so a
total of 41 videos will be shown at most. An indicator showing that there are
more elements will be added if more than the displayed videos are in the queue.

For now this seems like the simplest solution but could, in the future, be
replace with a virtualized list.