Generalize id annotations
Previously, some had the specific IDENTITY strategy, while others didn't. This
now uses AUTO on all the ids for consistency. This also helps with possibly
support for postgresql since it would NOT use IDENTITY, but SEQUENCE instead.

Additionally this now also makes sure that IDs are considered to be final and
not updatable.