Allow for plugins to register search shorthands
For now these search shorthands are basically replacements for specify query
values. So one could register a shorthand for `incomplete` and it would expand
to `(-Studio OR -ReleaseYear OR -Director)`. This is a very simple and basic
implementation for now and does not limit the queries in any way other than
being valid queries themselves.