Refactor plugin handling
This rewrites major parts of how plugins are managed and loaded, keeping most of
 the external plugin API untouched. This is to reduce internal dependencies
throughout the system as well as allowing reloading of plugins from file.

Loading was previously not possible as there was no reference from the loaded
plugin to the file it was loaded from. Introducing such a functionality alone
would result in a mess given the interdependencies of components at the time,
specifically the plugin manager and the plugin loader. This has now been cleaned
 up so that the loader no longer depends on the manager, but the other way
around. Thus the manager can now trigger the loading of a specific plugin file
to reload a specific plugin.

Another issue with the previous implementation was how references to metadata
were handled causing DB lookups on plugin execution for the metadata they might
want to read. Thus the execution environment now depended on the metadata
process. Since this was only necessary for for mapping metadata name to metadata
 id to then select the value from a map, it has been refactored to be assigned
to when the plugin gets enabled. This unfortunately resulted in some changes to
the API for plugins when using `video.set()` for metadata they did not specify
themselves. On the flip-side, the change results in a more cohesive API since
the read/write operations of metadata now are the same for owned and unowned
metadata as shown below:

```kotlin
val versionMetadata = numberRef("Version") // Referencing metadata from elsewhere
val myVersionMetadata = number("Custom Version", ASC, 1)

onUpdate(LATER) {
    it[myVersionMetadata] = it[versionMetadata]
}
```

The shorthands now also get removed when a plugin gets disabled, which
previously just stayed active until the application restarted.