Return video if updated on event
When there's an event generated on the client, we send this to the server as a
notification. So far, the client then did not get any info if anything actually
got updated or not. In some cases, we already just reloaded the video to make
sure we had an up-to-date version cached in the client.

Now we actually return the client a 200 or 304 depending on if the video was
actually updated as a result of the event. If it was, we include the new video
as well to avoid needing another request.