Fix lingering open transactions
Once a thumbnail or video stream is requested, spring internally opens a
transaction because `open-in-view` is enabled. This was enabled to allow for
lazy loading of data, for example thumbnail references when returning a video.
It was tried to manually create transactions but this obviously did not solve
the problem as the transaction was opened in the background to which we don't
have any direct influence on.

This now disables `open-in-view` and thus needs to make sure that we have a
transaction set up for most requests. Additionally, to avoid lazy loading
issues, this introduces some DTOs for certain models that load the necessary
data before being passed on to the response. As a result, thumbnails of videos
are no longer full structs, but just the ID now, as the remaining values were
not used anyway.