Refactor plugin loading to not rely on paths
When running as a boot jar, resources may not be resolvable to a path and will
instead throw an exception. We don't _really_ rely on the path, it was just a
convenient base interface that supports both plugins that we ship with and
plugins that are from external source. As this is no longer true, we now
refactor to create a more generic plugin source.