Replace jobrunr with quartz scheduler
While JobRunr did fit the needs, I had trouble finding a good way to comply with
 LGPL. While in theory the source is there and you could re-compile the whole
application, I wasn't quite sure how well this satisfied the license
requirements and as a result was never confident in pushing this to the main
branch.

Now, we use the quartz scheduler which is a bit more complicated to use but
handles the same use cases that jobrunr covered. Similarly, the jobs are not
persisted to the database for now. This should not be problematic as we have
cleanup jobs running that will take care of the failed tasks in due time.