Directly use kotlin jsr233
So far we relied upon ktsrunner to handle evaluation of the kotlin scripts.
However, because of usually lagging behind in kotlin version as well as some
classpath issues with spring, we now use the jsr223 script engine directly.
Since ktsrunner was already just a thin wrapper around the kotlin engine, not
much had to change.