Handle search for duration
Searching for a duration was previously not possible and simply resulted in a
server error "not implemented". This now adds support for filtering the videos
given an exact duration, e.g. `Length:"24m 12s"`, as well as rough duration,
e.g. `Length~"5m 10s"`. The latter is currently including videos withing 3 units
 (minutes if hours are specified, otherwise seconds), which is currently fixed
but will get support for adjusting later on.