Fix exception when enabling previously enabled plugin
When enabling a previously enabled plugin, we ran the event handlers for the
missing videos before the metadata had been assign their ID's. This obviously
causes the metadata assignment to fail. We need to enable the plugin
functionality before the event handlers get run, thus setting up the metadata
ids.