Use database-fresh plugin info when reloading plugin
When reloading a plugin from disk an exception would occur because of a
pessimistic locking problematic. This is because the plugin info we store in the
 plugin reference is not always exactly up to date. Thus, it is possible that
it's a version behind the one in the database and thus causing an exception as
the versions would be behind. Thus we make sure to load a fresh version of the
information before proceeding further.

This also removes some unnecessary saves that happen on load of the plugin. The
plugin will always be saved on add and thus a save beforehand is not necessary.