Allow plugins to register update events asynchronously
Events, such as an update event, currently had to be handled synchronously. Thus
 a user would have to wait possibly a long time for the plugin to finish it's
update. For many tasks this is fine as they will never take a long time finish.
However, there are some tasks that will always be slow and it's better for these
 to be run separately. The normal operation would continue and the videos would
get updated once the task(s) have been completed. This could be used for example
 to run OCR or other image recognition algorithms as they're usually quite
resource heavy.